requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.29.0
bcrypt==4.2.0
certifi==2024.8.30
cfgv==3.4.0
click==8.1.7
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.2
fastapi-cli==0.0.5
filelock==3.16.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
identify==2.6.1
idna==3.10
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
nodeenv==1.9.1
passlib==1.7.4
platformdirs==4.3.6
pre_commit==4.0.1
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
rich==13.9.2
shellingham==1.5.4
sniffio==1.3.1
sqladmin==0.20.0
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.40.0
typer==0.12.5
typing_extensions==4.12.2
uvicorn==0.32.0
uvloop==0.21.0
virtualenv==20.27.0
watchfiles==0.24.0
websockets==13.1
WTForms==3.1.2
